+++ src/image/image.interface.ts
@@ -1,6 +1,6 @@
 export interface ThyImageMeta {
     name?: string;
-    size?: number;
+    size?: string | number;
 }
 
 export interface ThyImageInfo extends ThyImageMeta {
@@ -23,3 +23,12 @@ export interface ThyImagePreviewOptions {
     zoom?: number;
     rotate?: number;
 }
+
+export interface ThyImagePreviewOperation {
+    icon: string;
+    name: string;
+    action: (image?: ThyImageInfo) => void;
+    type?: ThyImagePreviewOperationType;
+}
+
+export type ThyImagePreviewOperationType = 'zoom-out' | 'zoom-in' | 'rotate-right' | 'download';