+++ src/notify/examples/module.ts
@@ -1,9 +1,6 @@
-import { NgxTethysModule } from 'ngx-tethys';
-import { THY_NOTIFY_DEFAULT_OPTIONS } from 'ngx-tethys/notify';
-
+import { ThyNotifyModule } from 'ngx-tethys/notify';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-
 import { ThyNotifyBasicExampleComponent } from './basic/basic.component';
 import { ThyNotifyCustomHtmlExampleComponent } from './custom-html/custom-html.component';
 import { ThyNotifyDetailExampleComponent } from './detail/detail.component';
@@ -11,6 +8,9 @@ import { ThyNotifyHoverExampleComponent } from './hover/hover.component';
 import { ThyNotifyTypeExampleComponent } from './type/type.component';
 import { ThyNotifyPlacementExampleComponent } from './placement/placement.component';
 import { ThyNotifyDetailOperationExampleComponent } from './detail-operation/detail-operation.component';
+import { ThyButtonModule } from 'ngx-tethys/button';
+import { ThySpaceModule } from 'ngx-tethys';
+import { ThyNotifyCloseExampleComponent } from './close/close.component';
 
 const COMPONENTS = [
     ThyNotifyHoverExampleComponent,
@@ -19,21 +19,15 @@ const COMPONENTS = [
     ThyNotifyCustomHtmlExampleComponent,
     ThyNotifyBasicExampleComponent,
     ThyNotifyTypeExampleComponent,
-    ThyNotifyPlacementExampleComponent
+    ThyNotifyPlacementExampleComponent,
+    ThyNotifyCloseExampleComponent
 ];
 
 @NgModule({
     declarations: [...COMPONENTS],
     entryComponents: [...COMPONENTS],
-    imports: [CommonModule, NgxTethysModule],
+    imports: [CommonModule, ThyButtonModule, ThySpaceModule, ThyNotifyModule],
     exports: [...COMPONENTS],
-    providers: [
-        {
-            provide: THY_NOTIFY_DEFAULT_OPTIONS,
-            useValue: {
-                placement: 'topRight'
-            }
-        }
-    ]
+    providers: []
 })
 export class ThyNotifyExamplesModule {}